Fix CPU architecture code
Implement the CPU architecture code for macOS and Linux. Instead of
causing a name collision with SSE_VERSION, use the created bitmask to
both detect CMOV instruction and SSE2 instruction set and use in various
conditional statements.

Clean up naming to align with https://webkit.org/code-style-guidelines/
as we do not have any naming conventions place in this project.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>