Fix compiler warnings in Parameter.cpp
Set first byte of a string to zero instead of using the pattern
sprintf(prefix, "") as this will cause compiler warning on gcc and clang
and is also inefficient.

Fix a possible buffer overflow by increasing the prefix buffer size in
get_prefix() from 16 to 19 bytes so that it can fit the full integer
value in all cases.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>