app: Project template for a standalone app
Add a new target to premake5.lua called app that builds a standalone
application. Surge (and VSTGUI) files  are linked to the final
executable but it does not do anything useful yet. This just the initial
plumbing needed.

In addition, the target is only available on Linux. This is done
purpose. It is better to make it first work on one platform, and after
it is fully working, scale it to the other platforms. They are easy to
add once app itself is not a moving platform.

Linux is the most reasonable choice for this because on that platform
the standalone app is most desperately needed for development as plugin
targets do not work properly yet. Then, at least one target will support
Linux and probably the plugin issues get more easily fixed.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>