Correct windows.wt wavetable header
The file header indicates that there are sixteen waves contained in the
file when there are actually only nine.  This results in allocating more
memory than necessary, and attempting to read more data than the file
contains.

Modify the header to indicate the correct number of waves in the file.