Integrate SPD Logger (#94)
To allow us to have easy scalable logging available while developing and
for error reporting and the like, I integrated spdlog. See issue #94 for
further conversation on this decision. The logger initializes to
write to stderr and a file (on linux and mac) both and registers
a default logger at dll load time.