Update Windows compiler requirements and enable over-aligned allocations
Be explicit on minimum version for Visual Studio. 15.5 is a good entry
level version for us because it is the first version of Visual Studio
that has support for over-aligned allocations. Update premake5.lua to
enable this support. This is needed for SSE2 opcodes that require
16-byte aligned data.

For Linux and macOS we need not to do anything because they align all
allocations at least to 16 bytes.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>