Retina Bitmaps
If scalable bitmaps are present we can begin to implement zoom.
Step one is to implement Mac Retina display. We do that here
by loading 2x bitmaps if available and drawing them at 0.5x
scale. On retina displays this looks great. On regular displays
it is a no-op basically. See #226 for more.

None of this is activated if the skins aren't installed and
by default it falls back to the standard bitmaps.