Stubbing in start of Sketch Vectors.
This the start of the sketch file I have started. I wanted to replicate the orgional look as closely as possible. This provides a known state of usabilty to start from.

The organization of this file is not ready for prime time yet. I want to put this somewhere to ensure it doesnt get lost because it does represent a fair amount of work.

I have gone down the rabbit hole of recreating pixel by pixel much of the UI.