Throw SurgeError when configuration.xml is not found
Add a custom exception class SurgeError that is used to fail SurgeStorage
constructor when configuration.xml is not found. This is needed to avoid
null pointer dereferences because the code after that assumes that the
loading succeeded.

Scope this only for Linux now up until we have a functionality to display
an error dialog for all platforms. This commit prevents Renoise from
exhausting resources and finally crashing in the situation where
configuration.xml is not found.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>