Fix VST2 Learned Automation
This diff fixes automation in VST2 in Live, Bitwig, Reason etc...
The basic problem is that the display control ids differ from the logical
control ids. The synth remap does logical remapping but the display
remapping - which is implemented by "- start_paramtag" or "+ start_paramtag"
in the code doesn't flow to events uniformly. Specifically it didn't
flow to beginEdit and endEdit in VST2. So expose that "-" operation as
a well named function, override begin and end edit in vst2 host,
and voila.

We will definitely need to do the same with vst3 and au somehow.
Still figuring out those codepaths though.