VST2: Fail safely when the initialization fails
If the initialization fails, do not try to reinitialize Surge and free the
allocated memory when this happens. These were the two root causes that
caused Renoise on Linux to eat all memory and stall the whole DAW. Now the
plugin crashes but Renoise stays alive and responsive.

This commit introduces three states for the plugin: uninitialized,
initialized and dead. When the initialization fails the plugin is marked as
dead. All the machinery for dealing with states is baked inside the
function called tryInit() that is a rebaked version of the earlier function
that was called init(). It can be called in the beginning of any callback
as it does nothing unless the state is uninitialized, and returns a boolean
value telling whether the plugin has been initialized or not.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>