Fix two menu reset problems
As reported in #268, the menu reset i put was too heavy handed.
Make it so it cant reset "Timbre and the like".

As reported in #261, if you are in auto-name mode and remove
your first mapping you shoudl change the name. Comment this since
the logic required more than 10 seconds thought.