Focus issues in save
The vstgui and various mac hosts have wierd focus change issues
which don't update the value in some cases, which are both
host specific and wierd. Instead of fixing that underlying
scary issue, we can just put the editors in immediate mode
instead and get a working dialog.

This fixes the issue reported in #231