Original skin: Fixing Sketch Source file (#296)
* Stubbing in start of Sketch Vectors.

This the start of the sketch file I have started. I wanted to replicate the orgional look as closely as possible. This provides a known state of usabilty to start from.

The organization of this file is not ready for prime time yet. I want to put this somewhere to ensure it doesnt get lost because it does represent a fair amount of work.

I have gone down the rabbit hole of recreating pixel by pixel much of the UI.

* Revert "Stubbing in start of Sketch Vectors."

This reverts commit 4f43e5082f200f58be46eece5bb3b74f4f857161.

* Adding Sketch Source File for 1:1 of original skin

Adding Sketch Vector Source file. Also adding rendered sprite sheets.

I went a little nutty doing this. While this is very close the orginonal it lacks the brused metal and the splash screen isnt done.

* Fixing  Main Screen On Skin Renders

Was named bmp0012.png now correctly named bmp00102.png

* Fixing source file issues. Rerendering files. Converting Fonts to outlines

There were a couple outstanding issues with the last checkin. The big deal with this check in is that I converted the fonts to outlines to hopefully help with ledgibility.

* Fixing Filter F1 and F2 label

Getting label for F1 and F2 back into the skin. Fonts to outlines again.

* Fixing Sketch Source

Organizing Sketch file

Fixing issues with ADSR Env drawing. Slight stroke around warm, neutral, brigth buttons. Renders of BG again as well.

* Revert "Fixing Filter F1 and F2 label"

This reverts commit b1dae9c37243bd98aa417d01d2efa8e6e5dc964f.

* Revert "Fixing source file issues. Rerendering files. Converting Fonts to outlines"

This reverts commit ba5686712b5afa24366026967148f3a8caec9166.

* Revert "Fixing  Main Screen On Skin Renders"

This reverts commit 8c23f5d89c0749d00949c0010e2bb242c34a6237.

* Revert "Adding Sketch Source File for 1:1 of original skin"

This reverts commit 863edf128f4f68abbfb2ad5440e0b2269c6eb514.