Fix ordering when browsing through patches and categories
Add a field order to the classes patch_list and patch_category so that we
can move to the next patch based on the ordering.

Split SurgeSynthesizer::increasePatch() into increasePatch() and
increaseCategory() for clarity as the function contains essentially
contains two completely isolated code paths. In order to simplify the
implementation, allow the new functions to only move to the next and
previous patch and not an arbitrary number of patches.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>