Implement a User Reporting class (#301)
There are lots of times when we want to pop messages to users,
and as we find more cases where we want to tell users things, the
code was going to be littered with #if mac this #else windows that.

So make a namespace which has some basic functions for reporting an error
and doing an OK cancel; implement it on our platforms under the covers;
and use it in two places (configuration.xml missing and re-save a patch).

Get it working on Mac Windows and Linux. And voila

Also integrate @jsakkine review:

@jsakkine made an excellent point that there was no need for a class
or inheritance hierarchy when we could just have free functions in
namespaces. That's the way we want the future to look so lets start
that now.

Closes #217