Capitalize global constants (#308)
We can probably agree that a good convention for constants, defines and
enums is to have them capitalized. Right now it is hard to recognize from
source code, what is a constant and what is a normal variable.

There are still lot of enums etc. in the source code that are in lower
case but they are way more localized. This is a good place to start.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>