VST2: align the plugin class sanitize exception handling
Use alignas() to align synthetizers to 16 bytes, and sanitize the exception
handling in Vst2PluginInstance::tryInit(). Synthesizer and GUI editor are
first allocated to temporary variables inside a try-catch-block, and only
if everything goes well, they are assigned to the instance variables.
The instances are freed if something goes wrong.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>