Align text in orange boxes
This is not a pretty diff. But it is in the middle of some code which
is very explicit about layout and the layout was wrong. As shown in
issue #295, the combination of graphics, box placement, and label on the
lfo selector were just a bit off. This makes it almost not a bit off.
Ideally #214 will save the day in the future by letting us have a uniform
font across all platforms.