Implement a settings menu
We used to have the 'about' button but that only does one thing
As we want to do more ot the synth, we want options, and that's
a great place to hang them. So change it to 'menu' and make it pop
up a menu with a couple of options.