Remove CpuArchitecture
Only QuadFilterUnit and WindowOscillator check CaSSE2. None of the code
base uses CaCMOV to check anything. We can conclude from this that the
flags are useless because in practice they are not used at all.

CMOV instructions were introduced in 1995 with P6 microarchitecture (from
Pentium Pro to Pentium III) and SSE2 was introduced in 2000 with P68
microarchitecture (Pentium 4). I think the right thing to do is to remove
the checks. Even the most conservative studio environments have newer
hardware. Most of the operating systems would not even boot properly with
hardware older than Pentium 4.

Fixes: #350 ("CpuArchitecture problems")
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>