Limit zoom by screen
The current implementation will let you zoom to a size much larger than your
screen, a change which you can't undo. Reported in #369.

To resolve this, query the screen size at zoom time. Querying the screen
size means I added (in this diff) DisplayInfo.h which is an implementation
of functions in Surge::GUI namespace to get backing scale and screen size
for the screen in which your current frame is resident (or if you are split
screen, the screen the OS determines is the actually 'screen' for your
window). Implemented in same cross platform style as UserInteractions.

We then call that with a check and clear error message in SurgeGUIEditor

Incorporate @jsakkine review comments