Documents Patches with SubFolders
Surge loads patches from several documents folders in a single
function and uses the directory name as a category. Prior to this
change, only one level directory was used and that was the single
category name. So "<dir>/foo/x.fxp" made patch x in category foo.
This change recurses directories so "<dir>/foo/bar/x.fxp" makes patch
x have category "foo/bar" which shows up as a sub-menu in the
UI (even though the category is sitll a single flat category named
"foo/bar"). This works at retrive time and also at save time (so
saving a category "foo/bar" creates a subdir and so on). Addresses
part of issue #288