VST2 Zoom Mac/Win and Vector Assets on Win
* Implement zoom for VST2 Mac and Windows

This diff implements zoom for VST2 Mac and Windows using the
VSTGUI scale APIs in conjunction with the existing ability
of surge to register a zoom callback.

This would have been a relatively prosaic change were it not for
two bugs in VSTGUI

1: When you setzoom on a frame, it does not setzoom on the frame->getBackground
and so we need to explicitly resize the background image in VST2 as we zoom.
See the instance issues and additional scale in CScalableBitmap and SurgeBitmaps

2: the frame->getCurrentMouseLocation method is mistakenly scaled up twice
(as opposed to, one supposes, scaling up and down). So we need to, for vst2,
scale down twice.

This is all discussed in github issue #356.

With these diffs, though, both mac and windows vst2 plugins support
zoom (even though, ad this time, windows is using original not vector
bitmaps).

* Load Vector Assets on Windows

Vector Assets on windows loaded by automatically generating the
.rc file and associated assets needed for this (somewhat tedious)
enumeration.

See the extensive comment in scripts/win/emit-vector-rc.py for
details of the strategy.

This diff also includes feedback from @jsakkine which includes:
expand documentation, rename some unclear functions and variables,
and correct some comment inconsistencies. Also fix some spacing
inconsistencies.