Linux Build for VST24
Alas, the VST24 library requires a definition of __cdecl on linux.
There's a few ways to work around this, including editing the library
or defining it on a where-included basis, but the smallest approach
is to define the symbol __cdecl for vst24 on linux in the build pipeline.

The file which mis-requires it is aeffect.h which defined VSTCALLBACK
as __cdecl. Since VSTCALLBACK mis-definition is used in vst3sdk also
a source-only fix is prohibitve.

While at it, also expand .gitignore and make build-linux clean-all
more completely