Update to Jan 2019 vst3sdk (#401)
This substantial diff undertakes the mechanical changes to update
to the Jan 2019 master version of vst3sdk (hash 82380a8f). This
move consuming nearly 7 months of changes introduced several breaking
changes in our code which required modification, and introduced a couple of
new dependencies and links.

The primary thing breaking the builds was that most classes, including
common ones like CRect and CPoint, are now in the VSTGUI namespace.
This diff follows the convention of never using a namespace in a header
file, and thus has quite a few changes which are namespace additions in
headers and using statements in cpp files. But ironically, since my first
attempt was to just put using in a header and break the rule for once,
if you put using in the header, the code doesn't compile on windows
since VSTGUI::IDropTarget and ::IDropTarget collide. So has to be this way.

On linux several xcb and xkb dependencies emerge with this vst, so they
were added to the azure pipeline. Also the vstgui_linux.cpp is no longer
needed in the way we link, and if linked, doesn't work.

On windows, an ambiguity in std::numeric_limits<T>::max() requires us to
set -DNOMINMAX=1 for the time being, alas.

On mac, we expand the number of errors reported to 'as many as you generate'.

Commands used to update the module were

 git checkout master
 git checkout -b vst3sdk-update-373
 git submodule update --remote --merge --recursive
 git add vst3sdk/
 git commit
 git submodule update --init --recursive

This commit addresses github issue #373.