Enable tooltips
VSTGUI has pretty good tooltip support which is easy enough
to use. As described in issues #384 and #377 we can use these
immediately in two places. (1) when mousing over a slider to
show the current value and (2) when mousing over the patch
browser to display the otherwise unused patch comment.
This diff accomplishes that bu calling the appropriate
setAttribute call at the right times in the value change
and patch set codepaths.

Thanks to @esaruoho for spotting an error in the rebase.