No more using namespace; win -> c++17
using namespace in header files is bad for a variety of
well documented reasons. This change removes them in favor
of (1) full scoping in header files and (2) moving the
using namespace std to cpp files after the last include

The reason to do this change is so we can compile windows
with /std:c++17 which this change also does. Without this
change, the using namespace std would conflict between
byte and std::byte in system headers. However, with this
change, combined with the NOMINMAX requires by the vst3sdk
upgrade, min( and max( in our copy of wtl are ambiguous. As
such I took the distasteful route of modifying lib/wtl to
add `#include <algorithm>` and use std::min and std::max.

With this diff, both mac and windows compile and run and
windows compiles with C++ 17

A prior change has resulted in linux linking problems which
I'll resolve separately, but this code all compiles to objects
on linux.

Finally modify the hashbang of build-osx to help @kzantow
debug why this didn't azure build in the pipeline.

Move to /bin/bash