Iniclude lipng for windows in teh repository, rather than using nuget
Nuget turned out to be very unreliable, as it isn't included by default by visual studio 2017 and alhough it and MSBuild understands the nuget syntax, it can't be coaersed into resotring the packages as part of the build process.

I also changed the build.cmd script to call git submodule update, so that a user doesn't have to do that manually to make sure the submodules are in sync.

Also enabled /bigobj compile option on windows, as I could no longer build the debug x64 build without it.