Restore RMB for category patches
The RMB in category patch browser shows just that category. With
category menu refactoring, this behavior was temporarily lost.
With current code structure, we need to select a category
in the init state if you do RMB, or not show the menu. I decided
to do a scan for the init category and, barring that, use
category 0.

Closes #430