Resource manager for bitmaps on Linux
Piggy pack bitmap files to a blob and link in to the plugin binary.
Implement findMemoryBitmap() to look them up in CScalableBitmap and use
IPlatformBitmap::createFromMemory() to load them.

Cleaned up the build-linux.sh from some inconsistencies while doing this
i.e.  '--install' should actually be '--install-local'. There should
eventually be a separate '--install' option used for packaging that deploys
files under /usr/share. The local install for is for development purposes.
Removed '--build-and-install' as it is just as easy to chain commands with
'&&'.