Replace __linux__ with LINUX
Make the Linux compilation flag more inline how the flags for macOS and
Windows have been named. Someone might say that, why don't you pick UNIX to
make code more compatible with CMake, which we plan to eventually to
migrate. There is a predefined flag called APPLE on CMake for macOS but
also UNIX is true on macOS compilation.

What people tend to do in their CMakeLists.txt usually is:

if(UNIX AND NOT APPLE)
    set(LINUX TRUE)
endif()

So, in a distant way, this commit makes also the code more ready for the
eventual CMake migration.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>