Replace ->ToElement with dynamic cast
The TiXmlDocument thing does a check in a member function of
"if(this &&" which is non standard. Windows is fine, mac warns.
But gcc -O3 correctly optimizes it away. That optimization means
that linux release throws SEGV all over the place. Replace the
faulty if(this with a utility macro that actually test for null
outside the this instance.

See issue #469 for a more fulsome conversation of this change.

iMove to a macro