Shrinks below 100% in VST-en
VSTGUI has all sorts of background image scaling bugs. When you
grow the background doesn't. When you shrink, it also doesn't.
But the bounding box does. So back out the fact that we need
to grow it > 100 by not double shrinking it <100 and add with
another if and another comment.