A Variety of Zoom Enhancements
Once folks started using zoom, a variety of small enhancement
requests came in. This PR covers all of them, covering several
tickets.

The issues resolved are:

Closes #483.
Zoom implemented on linux; in doing so remove the HOST_SUPPORTS_ZOOM
switch, since all hosts support zoom. To make it testable, ahead
of implementing DisplayInfo on linux, set the defaulted screen size
to 1400x1050, a mid-range 15" laptop.

Closes #433
Correct the error message at zoom time to explain we add
a window-decoration-protection factor but don't show the
confusing dimensions that result.

Closes #473
Order and add separator to the setting menu to be more
rational. Keep "About" at the end.

Closes #457
If you ask for a zoom which outstrips screen parameters, rather than
retain the current zoom look for the largest zoom that will fit those
paramemters. Don't display the confusing-to-users message about how
much space we are reserving for window decoration.