Headless Surge working
The gui is well separable from the patch/parameter/dsp/synth
code so a headless mode should have been readily doable at
any time. With this commit, it is done. This shows a small
example of setting up and running a 2 oscillator detuned saw
configuration of surge.

On Linux, one VSTGUI symbol is still pulled in for a reason
I can't determine, so introduce it in a fixes file which we
need to address in a future issue if/as we develop headless.

Integrates @jsakkine review comments.

Closes #499.