Fix Linux Drawing Errors
Linux drawing was mis-rendering terribly. Turns out this was
because, for some reason, the inbound alpha to bitmap::draw
was 255, not 1.0f, and cairocontext.cpp has a branch decision
based on equality with 1.0f. This indicates there may still be
a problem in vstgui (why was the 255 inbound?) but an interim
fix on linux is to reset alpha=255 to alpha=1.0 which completely
fixes the painting issues reported in #513

Closes #513 - Linux Painting Issues