A Proposal for organizing Headless
This PR expands the capability of the headless app to start adding
functions and helpers we can use to debug. Specifically this adds
HeadlessApps.h and HeadlessTools.h which have various top level
activities and tools you can use, and replaces the main (which is
still available as Surge::Headless::runSimpleSynthToStdout) with
a program which loads all the presets into a headless surge.

Soon I'll make it so it can play them too, probably by adding
a "HeadlessPlayer.h/cpp" which plays various note sequences
into various synths.