Use Lato on MacOS and Windows
As described in #214, we want to use the OFL-licensed font
Lato for all components in Surge with the constraint that we
want to ship the font with the DLL and not require it to be installed
on an end users system.

To do this cross platforms, introduce a new C++ file
"RuntimeFont.h" which defines Surge::GUI::initializeRuntimeFontFromDistribution
and implement that in src/mac/RuntimeFontMac.cpp, Windows, etc.
This PR implements for Mac and Win and neither implements
nor references the symbol in Linux.

This change implements that in macOS by doing the following

1: Introducing the font to resources/fonts
2: Moving the fonts to the bundle with scripts/macOS/package*
3: Moving the font initiation for minifont and patchfont to
   runtime rather than surge dll load time (but only do this
   for fonts which USE_RUNTIME_LOADED_FONTS, which right now
   is only MAC. Once all platforms are done, that switch will
   disappear)
4: Write the MacOS specific CoreText code to load a font from
   memory into the system

This change implements that in Windows by doing the following

1: Including the font in the .rc file and subsequent res bundle
as a binary
2: Load the font and place it in a well known temp file
3: Load that temp file into the font registry using AddFontResource
4: Following the subsequent standard cfont implementation.