Two Zoom Changes (Resolution and Extrema)
Activate the 125x assets by making them callable in CScalableBitmap,
activating them in win script, and re-rendering the win resource files.

Also add a function "Zoom to Largest" which takes zoom to exactly the
point below where we would warn that you can't zoom any larger for your
screen, and "Zoom to Smallest", so that @sense-amr can do crazy things
with bitwig and Orca.

Moreover on smaller monitors, 95% zoom in logic pro is too large
and eats the menu button due to decoration, so in this change
we move to 90% of screen as max for all oses.

Closes 540: Activate 125% assets
Closes 590: Zoom to Biggest Menu