UTF8 for Wavetables; Noisier Mis-installation; Win Categories
UTF8 filename conversion is required on windows for vstgui so factor
the UTF8 string conversion into a function and adjust the callpoints.

Those callpoints intersect with places where we load patches and
wavetables, so at the same time, make the system report user errors
if wavetables or factory patches are non-loadable

In the course of testing this, I also realized that nested directories
didn't work properly on windows due to path separator being "\" rather
than "/", which this commit also addresses.

Closes #584 UTF8 for wavetables
Closes #543 Alert user to mis-installation

pathsetp