UserDefaults; Zoom Defaulting (#620)
Introduce a UserDefaults mechanism for surge, where users can set
a persistent value for some configuration element of surge and
store it between sessions. This is accomplished using the API in
src/common/UserDefaults.h which is pretty simple key/value store
and retrieve of strings and ints (now).

Use UserDefaults to create a default zoom option. This option
stores the defaultZoom and restores it across host and session
when you restart surge.

Closes #375: Have UserDefaults
Closes #374: Retain Zoom Settings
Closes #456: Allows a default zoom setting not 100%
Closes #586: Zoom state when open and close in session