Refactor Menus
This change is a large code change for zero functionality change,
required because of the semantics of menus and ::popup on Linux.
On Linux, because of various event loop handling, making
menu::popup() block is very difficult. The same idle loop which
pops up the UI also paints the on-screen menus, unlike in mac and
windows where they are separate elements.

So we have to change all of our menus from return-a-value-after-block
to bind-a-lambda-and-call fashion. That is, from

  id = 3;
  menu->add("do this", id );
  menu->popup();
  r = menu->getLastResult();
  if (r == 3)
    doThis()

we have

   menu>-add("do this", []() { doThis(); } )
   menu->popup()

which is cleaner but requires us to bind the correct state into lambdas
and reorder the code. Thus this commit is a massive code shuffle to
do exactly that, along with a helper function which actually adds a
no arg lambda to a menu in SurgeGUIEditor.

All of the menus were accordingly adjusted for copy, paste, rebind,
about, and so on.

Closes #536 popup menus on Linux