Defer bounds check on MacOS Default Zoom
On MacOS the bounds check happens before the window opens now
we have implemented default zoom; And the window is used to find
the screen for resolution check, so the resolution check fails on
large screens causing persisted large zooms to nonsensically reset.
Fix that by having the display size in the event of no window be
0,0 on macos (and 0,0 with no display on linux); supress the zoom
check size in that case; and check it again once the window is open
by forcing the idle loop by resetting zoom invalidation.