VST3 on Linux packages and installs
VST3 on Linux requires a mac-like bundle, as described in
the VST3 documentation vst3sdk/doc/vstinterfaces/vst3loc.html.
These changes move VST3 from a raw dll to a bundle using the
same techniques we use on macOS. Since we bake the images
and so on into the bundle, the bundle contains only the DLL.

With this change, bitwig studio will load our VST3, but it
crashes if you open the UI and gives corrupted parameter names.
But we are able to begin to debug.

Addresses part of #514