Reformat the AU Layer code
The AU Layer code was a mis-mash of tabs, spaces, styles, etc...
since it was never really cleaned up after being resurrected.
Since I have to go back in there for a few issues this was painful
so clean it up once.

This is a big textual diff but contains nothing other than formatting
and dead code removal.

Closes #663 Code Reformat AU