Plus key for zoom on Linux
Unsatisfyingly, the plus key (shift-= on US keyboards) on linux
doesn't report the character '+' to onKeyDown in VSTGUI but rather
reports '=' with shift depressed. Ideally the keyboard layer would
handle keyboard mapping of course (it does on mac) but this is a
useful enough feature and common enough keyboard mapping that I
choose to make the ugly choice to also bind shift-= to zoom increase
which defacto fixes the +/- zoom on linux.

Closes #600: +/- key on linux zoom