Remove the "app" (#686)
At the outset of surge development, we considered building a standalone
app versio of surge. Single-vst hosts on all platforms have made that
not necessary. Rather we have a headless version which doesn't load
or couple to the UI. So kill this vestigal app and the build thereof
which did nothing.

Closes #90 standalone surge