A More Expansive Headless Proposal
This commit suggests a way to implement Headless which is as a
player of internal events to buffers. For now the buffers are
vectors as are the events. Right now, this implements the
play and monitor of every patch with a simple C major scale,
as well as a quick test of a direct patch configuration.

This version includes changes after a review from @jsakkine