Set frame size when zooming (#713)
CFrame::setZoom sets the size of the frame to the ostensible
size, but if you zoom in and out a lot, it drifts by a pixel
here and a pixel there. One of those "(int)((int)a * 1.5)/1.5)"
over and over type things. Look in cframe ::setZoom and you will see.

But if after zooming you explicitly set the pixel size you want
this rounding leak disappears.

Closes #712 zoom size artifacts