VST3 Velocity
In VST3 NoteOnEvent::velocity is a float between 0 and 1.
Why? Who knows. But in every other midi thing it is a
char between 0 and 127. So when passed to surge directly
in VST3 it doesn't work. This change simply multiplies the velocity
and casts it to a char.

addresses parts of issues #389 and #26 but not all of them