Delay/Chorus menu fix
The reason the chorus-with-delay choruses were in the Delay/Chorus
menu was because the parent menu picked the effect type. So change
that so a snapshot can optionally have an 'i="num"' parameter which
overrides the type of the parent; and then move the Delay/Chorus
chorus settings into the Chorus menu; keep only two of them; rename
them to "Chorus (using Delay)" and "Wide Chorus (using Delay)"
and rename the awful "Delay/Chorus" menu to "Delay".

Closes #675