A small improvement to Stress (#796)
Stress has a pair of features. When testing for FAULT it tests
the RMS vs the peak as opposed to absolute PEAK to stop super loud
patches triggering a fake fault; and when a fault is found, it creates
a wav file so you can investigate the resulting wav form.