Move the QFC Init (#804)
I had implmented initToZero as a mebmer of QuadFilterChain but
QuadFilterChain is align malloced in a wierd way. While this
hadn't caused a problem yet, it gave me concern that the additional
vtable element would misalign the FBQ _aligned_malloc array
and so I moved it to a standalone function.