Allow channel aftertouch to modulate scene parameters (#818)
Channel aftertouch (as opposed to poly aftertouch) is a scene
level item in both MPE and non-MPE mode. (in MPE mode it is ignored
except on the global channel). As such it should be able to modulate
scene parameters. This change allows that by removing it from the
isSceneLevel exception list.

Closes #814