Alternate Tunings for Surge (alpha)
As part of the feature request in #828, this begins to add
alternate non-uniform tunings to surge. This commit is a good
step along the way to the final feature set but is incomplete.

With this commit, you can return a surge to an .scl file of your
chosing, but you cannot load a kbm, cannot choose a center note
other than C4/261hz, and the scl file you choose does not save
into either the patch or the DAW and is transient.

I am adding this commit, though, so that (1) users more experienced
with tuning can test this in the nightlies, (2) I don't have to
keep rebasing it forward and (3) I am guaranteed to get this feature
properly into 1.6.2