Correct Mac Shift-Wheel gesture (#834)
As discussed in #831, on mac shift-wheel (or shift-two-finger-drag
on trackpad) went the wrong way for horizontal actions. This
was due to vstgui handling a condition from old versions of macos
which are not present in current versions. Because of the nature of
the condition, we had to fix it by patching our vstgui.surge which
this PR does.