Implement FileOpen dialog in UserInteractions (#839)
There are increasingly cases where we want to prompt users for a file.
Rather than write this in VSTGUI I'm using the OS specific APIs behind
a UserInteraction common point. As of this writing Mac and Windows work but
ignore the options for starting directory and filter; and Linux returns a
UserError.

Since this move required me to move to cocoa for Mac UserInteractions,
also set up the .clang-format file to work properly with objective c

Closes #829 (but will require us to open another issue for linux and the
undone parts of the API). Pushing this now so I can use it for scale file
opening in the tuning issue.