Add a "TARGET_RACK" around assembly
Some of the older DSP code in mac contains explicit assembly
code. VCVRack builds with gcc so this syntax doesn't work;
but there are equiavlent C++ builtins which compile on linux.
So add a TARGET_RACK which is defined nowhere in surge project
itself which is used to condition out the assembly on all platforms,
not just linux.

Partly addresses #849 but this code should be checked to see if it is
even used and stuff still.