Unsatisfyingly handle mis-behaving mac hosts (#871)
Some mac hosts use their bundle, not the VST bundle, as
the bundle reference for the VST. Notably users report
Live 9.7 appears to do this. There's nothing we can really
do except do a hack and hope you installed in the global
location before bailing. But we might as well try that.

Closes #870