Add ARM NEON intrinsics
Use "-mfpu=neon" instead of "-msse2" when compiling for ARM.

In order to run Surge on ARM (e.g. Raspberry Pi 3) we need to support
NEON intrinsics. The fastest way to get this done right now is to use
https://github.com/jratcliff63367/sse2neon. This will also allow us to
maintain only one version of the DSP code. We use a copy of SSE2NEON.h
as using a submodule would be an overkill here.

Closes: #570