Null terminate ConfigurationXml.S
ConfigurationXml.S included a string but not a string range nor
a null terminator. We were mostly lucky that in surge this didn't
matter but with the onset of SVG and the build order in headless
we had SVG assets abut the XML configuration in a way which made
a non-parseable XML document. So just add a 0 into the rodata
segment right after the content. Fixes a crash in headless on linux.