Handle (kinda) zoom errors in Cakewalk
Cakewalk by Bandlab interacts with the VST3 zoom API in a way
which doesn't match the documentation in the VST3SDK; after calling
resizeFrame it calls back onSize() with a smaller frame even though
canResize is true; and even though it has grown its window.
So to avoid broken UIs disable Zoom in this host in Windows.

Closes #816