Correct the "Jump-to-zoom" startup
The initia zoom dance (show at 100 then go to the zoom you want)
was particularly annoying but, because of the way that VST2 and 3
differe in handling zoom, a little annoying to fix. These changes
make sure that the right size is clobbered onto the frame
at at least the right time, if not more, resulting in all of
AU, VST2 and VST3 opening at your default zoom cleanly.

CLoses #844