Increase patch display invalidation aggression
The patch display updates when queue_refresh is true but some
users report that it doesn't change when you pudate a patch.
That could happen if there is a race where the idle thread
runs ahead of the patch load thread, using the old version
but also resetting the refresh state. So add a third refresh
state which is "the patch browser id is not the synth id" which
forces a redraw.

May close the issue #817 but need user feedback.